+++ src/generators/model/default/model.php
@@ -76,7 +76,7 @@ class <?= $className ?> extends <?= '\\' . ltrim($generator->baseClass, '\\') . 
 <?php foreach ($relations as $name => $relation): ?>
 
     /**
-     * @return \yii\db\ActiveQuery
+     * @return <?= $relationsClassHints[$name] ?>
      */
     public function get<?= $name ?>()
     {