+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii Framework 2 gii extension Change Log
 2.1.4 under development
 -----------------------
 
-- no changes in this release.
+- Enh #287: Model generator is now generating relation's phpdoc hints with target ActiveQuery class (bscheshirwork)
 
 
 2.1.3 November 19, 2019